docker-deploy/web-app/requirements.txt
psycopg2
Django~=4.0.1
Pillow~=9.0.0
django-crispy-forms
